+++ gd-frontend/src/components/date-preview.js
@@ -14,7 +14,7 @@ class DatePreview extends Component {
         return (
             <div className="date-preview">
 
-                <li className="child datePreview" key={this.props.id}>
+                <li className="grid-child datePreview" key={this.props.id}>
                     <Link to={`/activities/${this.props.id}`}>
                         <img className="thumbnails" src={this.props.image} />
                         <p>{this.props.title}</p>
